+++ webpack.config.js
@@ -91,8 +91,18 @@ outputs.forEach(info => {
       setImmediate: false
     },
     externals: {
-      'bitcore-message': 'bitcoreMessage',
-      'node-forge': 'forge',
+      'bitcore-message': {
+        amd: 'bitcore-message',
+        commonjs: 'bitcore-message',
+        commonjs2: 'bitcore-message',
+        root: 'bitcoreMessage'
+      },
+      'node-forge': {
+        amd: 'node-forge',
+        commonjs: 'node-forge',
+        commonjs2: 'node-forge',
+        root: 'forge'
+      },
       jsonld: 'jsonld'
     }
   };